New peers are added to peer list and db
When new peers are posted to the endpoint they are now added to the DB
and the networking interface is also on the fly configured to have the
new peer in it's list.

Still need to test if a interface restart is required before packets can
be accepted from the new peer.